Api: Create models' super class (fate#323437)
