+++ hawk/app/models/api/v1/resource.rb
@@ -1,10 +1,6 @@
-require 'cibtools'
-require 'rexml/document' unless defined? REXML::Document
-require 'rexml/xpath' unless defined? REXML::XPath
-
 module Api
   module V1
-    class Resource
+    class Resource < Cib
 
       def initialize(root, id)
         @id = id
